Move config/headAssets > pages/assets
