+++ app/pages/assets.js
